+++ lib/NetworkService/Model/no_internet_model.dart
@@ -1,15 +1,18 @@
 import 'package:flutter/material.dart';
-
-/// widget : -  your no interenet widget if not passed default is set
-/// backgroundColor :- change color of default no internet widget
-/// displayImage : - change default icon to your image
-/// titleStyle :- change deafult style of title
-/// descriptionStyle :-  change default description style
 class NoInternetWidget {
+  /// widget : -  your no interenet widget if not passed default is set
   Widget? widget;
+
+  /// backgroundColor :- change color of default no internet widget
   Color? backgroundColor;
+
+  /// displayImage : - change default icon to your image
   Widget? displayImage;
+
+  /// titleStyle :- change deafult style of title
   TextStyle? titleStyle;
+
+  /// descriptionStyle :-  change default description style
   TextStyle? descriptionStyle;
 
   NoInternetWidget(