+++ lib/NetworkService/connectivity_watcher.dart
@@ -1,6 +1,6 @@
 export 'widgets/connectivity_watcher_app.dart';
 export 'widgets/connectivity_widget.dart';
 export 'widgets/default_no_internet.dart';
-export 'Model/ConnectivityWidgetModel.dart';
+export 'Model/no_internet_model.dart';
 export 'Model/connectivity_enum.dart';
 export 'utils/connectivity_watcher.dart';