+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Release
 * 0.0.1  initial release.
-* 0.0.2 Added Default No Internet widget and minor fixes 
-
+* 0.0.2  Added Default No Internet widget and minor fixes 
+* 1.0.0  Added ConnetivityAppWrapper and improved functionality